Update Models.py
Even though this file is very short, a lot is going on here. I am training two models, specifically a Bernoulli Naive Bayes classifier and a "One-Versus-One" support vector machine classifier. The data is imported from a CSV file constructed earlier, where the majority of the work is done to achieve the highest accuracy possible. The data is then shuffled for  more complete and distributed training and test sets, and then the data is split at around 70/30. My choice of using Bernoulli Naive Bayes here lies in the fact that the data itself is comprised of binary feature vectors, which Bernoulli Naive Bayes is explicitly meant to handle, as well as it's small size in general. The reported accuracy's of both models seem to be around 95% and 94.5%, respectively.